+++ api/src/user/service/user.service.ts
@@ -4,7 +4,12 @@ import { catchError, from, map, Observable, switchMap } from 'rxjs';
 import { AuthService } from 'src/auth/services/auth.service';
 import { Repository } from 'typeorm';
 import { UserEntity } from '../models/user.entity';
-import { User } from '../models/user.interface';
+import { User, UserRole } from '../models/user.interface';
+import {
+  paginate,
+  Pagination,
+  IPaginationOptions
+} from 'nestjs-typeorm-paginate';
 
 @Injectable()
 export class UserService {
@@ -23,7 +28,8 @@ export class UserService {
         newUser.password = passwordHash;
         newUser.username = user.username;
         newUser.id = user.id;
-        newUser.role = user.role;
+        // set users role to `user` by default
+        newUser.role = UserRole.USER;
 
         return from(this.userRepository.save(newUser)).pipe(
           map((user: User) => {
@@ -64,6 +70,17 @@ export class UserService {
     );
   }
 
+  paginateUsers(options: IPaginationOptions): Observable<Pagination<User>> {
+    return from(paginate<User>(this.userRepository, options)).pipe(
+      map((usersPagination: Pagination<User>) => {
+        usersPagination.items.forEach((user: User) => {
+          delete user.password;
+        });
+        return usersPagination;
+      })
+    );
+  }
+
   deleteOne(id: number): Observable<any> {
     return from(this.userRepository.delete(id));
   }
@@ -71,6 +88,7 @@ export class UserService {
   updateOne(id: number, user: User): Observable<any> {
     delete user.email;
     delete user.password;
+    delete user.role;
 
     return from(this.userRepository.update(id, user));
   }